Added support for C++ programs using the libst.

+++ src/avg.c
@@ -103,7 +103,7 @@ eff_t effp;
 void avg_flow(effp, ibuf, obuf, isamp, osamp)
 eff_t effp;
 LONG *ibuf, *obuf;
-int *isamp, *osamp;
+LONG *isamp, *osamp;
 {
 	avg_t avg = (avg_t) effp->priv;
 	int len, done;
