+++ src/copy.c
@@ -45,7 +45,7 @@ eff_t effp;
 void copy_flow(effp, ibuf, obuf, isamp, osamp)
 eff_t effp;
 LONG *ibuf, *obuf;
-int *isamp, *osamp;
+LONG *isamp, *osamp;
 {
 	int done;
