+++ src/deemphas.c
@@ -159,7 +159,7 @@ eff_t effp;
 void deemph_flow(effp, ibuf, obuf, isamp, osamp)
 eff_t effp;
 LONG *ibuf, *obuf;
-int *isamp, *osamp;
+LONG *isamp, *osamp;
 {
      deemph_t deemph = (deemph_t) effp->priv;
      int len, done;
@@ -182,7 +182,7 @@ int *isamp, *osamp;
 
 void deemph_drain(effp, obuf, osamp)
 LONG *obuf;
-int *osamp;
+LONG *osamp;
 {
      /* nothing to do */
 }
