+++ src/flanger.c
@@ -172,7 +172,7 @@ eff_t effp;
 void flanger_flow(effp, ibuf, obuf, isamp, osamp)
 eff_t effp;
 LONG *ibuf, *obuf;
-int *isamp, *osamp;
+LONG *isamp, *osamp;
 {
 	flanger_t flanger = (flanger_t) effp->priv;
 	int len, done;
@@ -208,7 +208,7 @@ int *isamp, *osamp;
 void flanger_drain(effp, obuf, osamp)
 eff_t effp;
 LONG *obuf;
-int *osamp;
+LONG *osamp;
 {
 	flanger_t flanger = (flanger_t) effp->priv;
 	int done;
